+++ virtual_network/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## aks_node_pool-v0.6.0
+
+### Updated
+- updated `azurerm` provider to `2.16.0`
+
 ## virtual_network-v0.5.0
 
 ### Updated